requirements.txt
django-ckeditor
south
mysql-python
Django
Markdown
PIL
PyYAML
django-filter
djangorestframework
